Update pydeck tooltip style and modularize the Jupyter widget tooltip (#3590)
